requirements.txt
# No external dependencies required - SQLite is built into Python
# Optional: for enhanced features
python-dotenv>=0.19.0